pyproject.toml
[project]
name = "planar-computer-use"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncvnc>=1.3.0",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "uvicorn[standard]>=0.34.2",
    "planar",
    "gradio-client>=1.10.1",
    "pillow>=11.2.1",
]

[dependency-groups]
dev = [
    "pyright[nodejs]>=1.1.401",
    "ruff>=0.11.11",
]


[tool.uv.sources]
planar = { path = "../planar", editable = true }
